custom-functions examples
